Push Auth Logic

+++ app/Http/Controllers/AuthController.php
@@ -15,18 +15,18 @@ class AuthController extends Controller
    public function register(RegisterRequest $request)
    {
     $validatedData=$request->validated();
-    $code = rand(1, 10000);
     $user=User::create([
         'name'=> $validatedData['name'] ,
         'email'=> $validatedData['email'] ,
         'password'=>Hash::make($validatedData['password']) ,
     ]);
 
-$data['token'] = $user->createToken('User')->plainTextToken;
-$data['name']=$user['name'];
-$data['email']=$user['email'];
+    $data['token'] = $user->createToken($request->userAgent() . ' - ' . now())->plainTextToken;
+    $data['name']=$user['name'];
+    $data['email']=$user['email'];
     return ApiResponse::sendResponse(201,'User Registered',$data);
    }
+
    public function login(LoginRequest $request)
    { 
     $validated =  $request ->validated() ;
@@ -38,29 +38,22 @@ $data['email']=$user['email'];
     
     if(!Hash::check($validated['password'],$user->password))
     {
-        return ApiResponse::sendResponse(404,'Incorect Paswword',[]);
+        return ApiResponse::sendResponse(403,'Incorrect Password',[]);
     }
-    $data['token']= $user->createToken('UserLogin')->plainTextToken ; 
+    $data['token']= $user->createToken($request->userAgent() . ' - ' . now())->plainTextToken ; 
     $data['name']=$user['name'];
     $data['email']=$user['email'];
     return ApiResponse::sendResponse(200,'User Logging in',$data);
    }
-//    public function logout(Request $request)
-//    {
-//     $token = $request->user()?->currentAccessToken();
-//      if($token)
-//      {
-//         $token->delete();
-//         return ApiResponse::sendResponse(200, 'Logged Out Successfully', []);
-//      }
 
-//      return ApiResponse::sendResponse(400, 'No active session found', []);
-
-//    }
  public function logout(Request $request)
     {
+    if(!$request->user()?->currentAccessToken())
+    {
+        return ApiResponse::sendResponse(400, 'No active session found', []);
+    }
     $request->user()->currentAccessToken()->delete();
-    return ApiResponse::sendResponse(200,'LOgout succsesfully', null);
+    return ApiResponse::sendResponse(200,'Logout successfully', null);
     }
